added openAI API library dependency in root
