+++ package.json
@@ -22,25 +22,24 @@
   },
   "homepage": "https://github.com/iskry/TalesfromtheAlgorithm#readme",
   "dependencies": {
-    "chatgpt-official": "^1.1.4",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.11",
     "@mui/styled-engine-sc": "^5.11.11",
-
     "@stripe/react-stripe-js": "^1.16.5",
     "apollo-client": "^2.6.10",
     "apollo-server": "^3.12.0",
     "apollo-server-express": "^3.12.0",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
+    "chatgpt-official": "^1.1.4",
     "concurrently": "^7.6.0",
+    "dotenv": "^16.0.3",
     "expect": "^29.4.3",
+    "express": "^4.18.2",
     "jest": "^29.4.3",
     "materialize-css": "^1.0.0",
     "mongoose": "^7.0.0",
+    "openai-api": "^1.3.1",
     "stripe": "^11.13.0",
     "styled-components": "^5.3.8"
   },
