+++ server/graphQL/index.js
@@ -1,14 +1,13 @@
-const { ApolloServer } = require("apollo-server");
 const { ApolloServerExPress } = require("apollo-server-express");
 const typeDefs = require("./typeDefs");
 const resolvers = require("./resolvers");
 
-const server = new ApolloServer({
+const apolloServerExpress = new ApolloServerExPress({
   typeDefs,
   resolvers,
 });
 
-server
+apolloServerExpress
   .listen()
   .then(({ url }) => {
     console.log(`🚀 Server ready at ${url}`);
@@ -16,3 +15,7 @@ server
   .catch((error) => {
     console.log("Error starting server: ", error);
   });
+
+module.exports = { typeDefs, resolvers };
+
+// tools of graphQL library (modules) exported to server file to be used by exprees host . the express host does its thing either at the view (tables) or the database (kitchen). Once the tools have been used on either client or server sided API, those tools return a new state and new rendered component
