+++ server/server.js
@@ -1,90 +1,39 @@
-const dotEnv = require('dotenv').config();
+const dotEnv = require("dotenv").config();
 const key = process.env.API_KEY;
 
+// mongoDB Database connect
+const mongoDB_connect = require("./config/db/connection.js");
+
+// graphQL tools import && then updating instance of apollo-server
+const { apolloServerExpress, typeDefs, resolvers } = require("./graphQL");
+const graphQL = apolloServerExpress({ typeDefs, resolvers });
+// takes tools out of work bag first before putting back "builds" once the app is "built" with initialized "api_state"
+
+// express
 const express = require("express");
-const mongoose = require("mongoose");
-
-const bodyParser = require("body-parser");
-
-const { ApolloServer } = require("apollo-server");
-
-
+// const router = require("express-router")
 const PORT = process.env.PORT || 3001;
-const app = express();
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-});
+const bodyParser = require("body-parser");
+const app = express()
+  //.router(); once routes are finished then can uncomment or else error will ensue
 
+// vroom_vroom-express-initialize
 app.use(bodyParser.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
+app.use(express.static("public"));
 
-app.get('/config', (req, res) => {
-  res.send({
-    publishableKey: process.env.STRIPE_PUBLISHABLE_KEY,
-  })
-})
-app.post("/create-payment-intent", async (req, res) => {
-  const paymentIntent = await Stripe.PaymentIntentsResource.create({
-    currency: 'usd',
-    amount: 1999,
-    automatic_payment_methods: {
-      enabled: true,
-    },
-  })
-  res.send({ clientSecret: paymentIntent.client_secret })
-})
-
-async function startServer() {
-  app.use(express.static("public"));
-  let options = {
-    temperature: 0.7,
-    max_tokens: 256,
-    top_p: 1,
-    frequency_penalty: 0,
-    presence_penalty: 0,
-    instructions: `You are ChatGPT, a large language model trained by OpenAI.`,
-    model: "text-davinci-003",
-    stop: "",
-  }; //can we modularize this out of the srver file as an idependent file or be in another existing file within the server directory eg. schema, model
-  const { default: ChatGPT } = await import("chatgpt-official");
-  let bot = new ChatGPT(key, options);
-  let response = await bot.ask(
-    "think of 5 funny and clever names about generative AI and storys. the AI will be fed prompts by the user and create custom stories. make it edgar allan poe themed. thanks "
-  );
-  console.log(response);
-  let conversationId2 = "another conversation name";
-  let response2 = await bot.ask("Hello?", conversationId2);
-  console.log(response2);
-}
-startServer().then (() => {
-  mongoose.connect(process.env.MONGODB_URI || "mongodb://127.0.0.1/algoDB", {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  }).then(() => {
-    console.log("Connected to MongoDB");
-    server.listen({ port: 5000 }).then((res) => {
-      console.log(`Server running at ${res.url}`);
-    });
-  });
-
-  app.listen(PORT, () => {
-    console.log(
-      `
-      ==============================
-      "Online at ${PORT}, Server is."
-                  __.-._
-                  '-._"7'
-                   /'.-c
-                   |  /T
-                  _)_/LI
-      ==============================
-
-    `);
-  });
+app.listen(PORT, () => {
+  mongoDB_connect();
+  graphQL.applyMiddleware({ app });
+  console.log(`
+    ==============================
+    "Online at ${PORT}, Server is."
+                __.-._
+                '-._"7'
+                  /'.-c
+                  |  /T
+                _)_/LI
+    ==============================
+  `);
 });
-
-
-
-module.exports = { server, Mutation };
